bookmark-manager/pyproject.toml
[project]
name = "bookmark-manager"
version = "0.1.0"
description = "Minimal bookmark manager with semantic search"
requires-python = ">=3.10"

[project.scripts]
bookmark = "cli.bookmark_cli:cli"

[project.optional-dependencies]
cli = [
    "httpx==0.27.2",
    "click==8.1.7",
    "rich==13.9.2",
]
server = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.31.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "sqlalchemy==2.0.35",
    "aiosqlite==0.20.0",
    "sqlite-vss==0.1.2",
    "sentence-transformers==3.2.0",
    "httpx==0.27.2",
    "jinja2==3.1.4",
    "feedparser==6.0.11",
]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
]
bot = [
    "python-telegram-bot==21.7",
    "httpx==0.27.2",
]

[tool.setuptools]
packages = ["cli"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
